Fix Delete Buttons and Timesheet Approval - Fixed the delete functionality of Allowance - Fixed the delete functionality of Government Contributions - Fixed the delete functionality of Employee Leaves along the deletion of related Filed Leaves - Fixed the delete functionality of Rates

+++ src/main/java/io/softwaregarage/hris/attendance/views/EmployeeShiftListView.java
@@ -232,8 +232,10 @@ public class EmployeeShiftListView extends VerticalLayout {
                         // Delete first the related timesheet of the employee.
                         List<EmployeeTimesheetDTO> employeeTimesheetDTOList = employeeTimesheetService
                                 .findByEmployeeDTO(selectedEmployeeShiftScheduleDTO.getEmployeeDTO());
-                        for (EmployeeTimesheetDTO employeeTimesheetDTO : employeeTimesheetDTOList) {
-                            employeeTimesheetService.delete(employeeTimesheetDTO);
+                        if (!employeeTimesheetDTOList.isEmpty()) {
+                            for (EmployeeTimesheetDTO employeeTimesheetDTO : employeeTimesheetDTOList) {
+                                employeeTimesheetService.delete(employeeTimesheetDTO);
+                            }
                         }
 
                         // Get the selected department and delete it.
